+++ core/task.cpp
@@ -712,8 +712,8 @@ QList<int> Task::uncompletedHarvestLinkedTasks(int plantingId) const
 
 QVariantMap Task::drawInfoMap(int taskId, int season, int year) const
 {
-    QElapsedTimer timer;
-    timer.start();
+    //    QElapsedTimer timer;
+    //    timer.start();
     const auto record = recordFromId("task_view", taskId);
     const auto assignedDate =
             QDate::fromString(record.value(QStringLiteral("assigned_date")).toString(), Qt::ISODate);
@@ -738,7 +738,7 @@ QVariantMap Task::drawInfoMap(int taskId, int season, int year) const
 
     //    qDebug() << description << graphStart << width;
 
-    qDebug() << "[drawPlantingMap]" << timer.elapsed() << "ms";
+    //    qDebug() << "[task info map]" << timer.elapsed() << "ms";
 
     return { { "graphStart", graphStart },
              { "width", width },